feat: update raised amount

+++ src/lib/model/dto/create-transaction.dto.ts
@@ -2,6 +2,6 @@ import { z } from "zod";
 
 export const transactionSchema = z.object({
     loanId: z.string().uuid(),
-    amount: z.coerce.number().int().positive(),
+    amount: z.coerce.number().positive(),
     type: z.string(),
 })
